Actualizacion

+++ src/main/java/com/reto3/reto3/Repositorio/MessageRepository.java
@@ -3,5 +3,5 @@ package com.reto3.reto3.Repositorio;
 import com.reto3.reto3.Entidad.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository extends JpaRepository<Message,Long> {
+public interface MessageRepository extends JpaRepository<Message, Integer> {
 }
