+++ src/main/java/com/reto3/reto3/Repositorio/ScoreRepository.java
@@ -3,5 +3,5 @@ package com.reto3.reto3.Repositorio;
 import com.reto3.reto3.Entidad.Score;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScoreRepository extends JpaRepository<Score, Long> {
+public interface ScoreRepository extends JpaRepository<Score, Integer> {
 }
